+++ src/data/import_from_affymetrix.py
@@ -175,10 +175,14 @@ def deal_with_report(report: str, dataset: Dataset, assembly: str):
     type=str,
     help="Source assembly imported_from",
     required=True)
+@click.option(
+    '--max_samples',
+    type=int,
+    help="Limit import to first samples (only valid for affymetrix report)")
 def main(
         prefix, report, dataset, coding, breed_code, chip_name, assembly,
         create_samples, sample_field, search_field, src_version,
-        src_imported_from):
+        src_imported_from, max_samples):
     """
     Read genotype data from affymetrix files and convert it
     to the desidered assembly version using Illumina TOP coding
@@ -239,7 +243,7 @@ def main(
 
     elif report:
         # this is an affymetrix reportfile
-        plinkio.read_reportfile()
+        plinkio.read_reportfile(n_samples=max_samples)
 
     # fetch coordinates relying assembly configuration. Mind affy probeset_id
     plinkio.fetch_coordinates(