+++ tests/data/test_import_from_affymetrix.py
@@ -259,6 +259,68 @@ class TestImportFromAffyReport(AffyTestMixin, unittest.TestCase):
             self.assertEqual(len(sample_list), 2)
             self.assertEqual(len(locus_list), 1)
 
+    @patch('src.features.smarterdb.Dataset.result_dir',
+           new_callable=PropertyMock)
+    @patch('src.features.smarterdb.Dataset.working_dir',
+           new_callable=PropertyMock)
+    def test_import_from_affymetrix_limit(self, my_working_dir, my_result_dir):
+        """Import from report by limiting the number of samples"""
+
+        # create a temporary directory using the context manager
+        with tempfile.TemporaryDirectory() as tmpdirname:
+            working_dir = pathlib.Path(tmpdirname)
+            results_dir = working_dir / "results"
+
+            # assign return value to mocked property
+            my_working_dir.return_value = working_dir
+            my_result_dir.return_value = results_dir
+
+            # copy test data files
+            self.link_files(working_dir)
+
+            result = self.runner.invoke(
+                import_from_affymetrix,
+                [
+                    "--dataset",
+                    "test.zip",
+                    "--report",
+                    "affyreport.txt",
+                    "--chip_name",
+                    self.chip_name,
+                    "--assembly",
+                    "OAR3",
+                    "--breed_code",
+                    "TEX",
+                    "--sample_field",
+                    "alias",
+                    "--create_samples",
+                    "--src_version",
+                    "Oar_v4.0",
+                    "--src_imported_from",
+                    "affymetrix",
+                    "--coding",
+                    "ab",
+                    "--max_samples",
+                    "1"
+                ]
+            )
+
+            self.assertEqual(0, result.exit_code, msg=result.exception)
+            self.assertEqual(SampleSheep.objects.count(), 1)
+
+            # check imported chip_name attribute
+            for sample in SampleSheep.objects:
+                self.assertEqual(sample.chip_name, self.chip_name)
+
+            plink_path = results_dir / "OAR3" / "affyreport_updated"
+            plink_file = plinkfile.open(str(plink_path))
+
+            sample_list = plink_file.get_samples()
+            locus_list = plink_file.get_loci()
+
+            self.assertEqual(len(sample_list), 1)
+            self.assertEqual(len(locus_list), 1)
+
 
 if __name__ == '__main__':
     unittest.main()