+++ HISTORY.rst
@@ -22,6 +22,7 @@ TODO
 0.4.7.dev0
 ----------
 
+* Import from affymetrix a limited number of samples
 * Skip sample creation when there's no alias
 * Support for missing columns in affymetrix report files
 * Support *invalid python names* in ``src.features.affymetrix.read_affymetrixRow``